Update ElasticSearch and Kibana to 5.4.2
This commit attempts to update logging stack to newer versions. The
images used are official docker.io ones. Update of ES required a newer
version of fluentd as well.

Upgrading fluentd resulted in hundereds of MBs of logs being printed in
fluentd docker right after it was up and running. The root cause of this
was multiline parser plugin which created infinite loops (pay attention
to nested 'got incomplete line before first line' log):
https://pastebin.com/1zsyEDN8 .
I've removed the part that caused this error as parsing multiline
tracebacks from Openstack services will be handled when the following
PS is merged:
https://review.openstack.org/#/c/490036/

Change-Id: I3e2c5c817f17dd344b56b717251afadc03e3f186
Co-Authored-By: Mateusz Blaszkowski <mateusz.blaszkowski@intel.com>