Elasticsearch: configuring log rotation
Before this change Elasticsearch logs were only printed to the console.
Now, every log is also stored in a file named the same as the hostName
of the instance, for example elasticsearch-data-0 pod prints its logs to
/var/lib/elasticsearch/data/elasticsearch-data-0.log file. This is done
due to the fact that logs from elastichsearch-data instances are stored
in the same persistent volume. When multiple instances were writing to the
same file, synchronization issues would appear (multiple processes
logging to the same file using RollingFile appender is not supported).

Rotation configuration:
- maximum number of 5 archived log files
- each log file can take up to 100MB
- when log file exceeds 100MB, it's renamed by adding an integer to the
end of the file, for example file.log is renamed to file.log.1

Change-Id: I4689661c72084ad2f9af94dc574208ae8e1856d7
Signed-off-by: Mateusz Blaszkowski <mateusz.blaszkowski@intel.com>