Update upgrade host playbook
playbook updated to match osh and match osh-infra

This updates the upgrade host playbook in openstack-helm to match the
playbook used in openstack-helm-infra. The recent addition of adding
an apparmor profile to the Calico chart requires us to do the same
setup on hosts in the openstack-helm jobs before attempting to deploy
Calico.

Change-Id: I0861038231e55a392b002d40476c40698d249260