add post-upgrade hook and deletion policy to readonly SA token secret
Before, the secret was not getting installed when upgrading from a
previous version of this chart without the secret. On clean install it
was working.

Now, with the addition of the post-upgrade hook this chart will install
the secret correctly when upgrading from a previous version without this
secret.

The deletion policy is required for sequential upgrades. Tiller
does not keep track of resources installed via post-{install,upgrade}
hooks, this causes an "Already exists" error when upgrading without this
hook. With this hook, the secret will be deleted before upgrading,
which will then install this resource.


Change-Id: Ia3af5af8bcf28cae3ad31f427068a025a5a4c7fd