Point to OSH-images images
We now have a process for OSH-images image building,
using Zuul, so we should point the images by default to those
images, instead of pointing to stale images.

Without this, the osh-images build process is completely not
in use (and completely opaque to deployers), and updating the
osh-images process or patching its code has no impact on OSH-addons.

This should fix it.

Change-Id: I23d69b56e6de4f0f76d6031b301e81a294ddcc50