fix(sonobuoy): only add apparmor for results-publisher when publishing
Before, if conf.publish_results was false then Helm would fail to
install the Sonobuoy chart because the apparmor annotation was being
added for the results-publisher container, but the container didn't
actually exist because it was disabled.

Now, the apparmor annotation is only included for the results-publisher
container when conf.publish_results is true.

Change-Id: I731b7d03c9699db0fcab61439479796617ebff2a